+++ modules/azure/github-runner/main.tf
@@ -15,7 +15,7 @@ terraform {
 
   required_providers {
     azurerm = {
-      version = "2.82.0"
+      version = "2.99.0"
       source  = "hashicorp/azurerm"
     }
     tls = {