+++ modules/kubernetes/fluxcd-v2-azdo/main.tf
@@ -21,11 +21,11 @@ terraform {
   required_providers {
     helm = {
       source  = "hashicorp/helm"
-      version = "2.3.0"
+      version = "2.4.1"
     }
     flux = {
       source  = "fluxcd/flux"
-      version = "0.5.1"
+      version = "0.11.2"
     }
     azuredevops = {
       source  = "xenitab/azuredevops"
@@ -33,11 +33,11 @@ terraform {
     }
     kubernetes = {
       source  = "hashicorp/kubernetes"
-      version = "2.6.1"
+      version = "2.8.0"
     }
     kubectl = {
       source  = "gavinbunney/kubectl"
-      version = "1.13.0"
+      version = "1.13.1"
     }
   }
 }