+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - [#593](https://github.com/XenitAB/terraform-modules/pull/593) Change Prometheus remote write settings to align with best practices.
 - [#574](https://github.com/XenitAB/terraform-modules/pull/574) [Breaking] Update cert-manager version.
+- [#595](https://github.com/XenitAB/terraform-modules/pull/595) Update Terraform provider versions.
 
 ### Removed